Split mapping files into individual files
